// types/navigation-types/stack-types/home-stack.type.ts
export type HomeStackParamList = {
  carsi_screen: undefined;
  hal_screen: undefined;
  pazar_screen: undefined;
  product_detail_screen: {
    productName: string;
    id: string;
  };
  hal_product_detail_screen: {
    productName: string;
    id: string;
  };
  pazar_product_detail_screen: {
    productName: string;
    id: string;
  };
};